examples/requirements.txt
torch>=2.0.0
lightning>=2.0.0
mlflow>=2.10.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0.0
pyarrow>=14.0.1
optuna>=3.4.0
plotly>=5.14.1